refactor: put simple functions used once in the calling file (2/3)
