+++ utils/make-single-node-comment-after-fixes.ts
@@ -1,8 +1,6 @@
 import type { TSESLint } from '@typescript-eslint/utils'
 import type { TSESTree } from '@typescript-eslint/types'
 
-import { getCommentAfter } from './get-comment-after'
-
 interface MakeSingleNodeCommentAfterFixesParameters {
   sortedNode: TSESTree.Token | TSESTree.Node
   node: TSESTree.Token | TSESTree.Node
@@ -43,4 +41,24 @@ export let makeSingleNodeCommentAfterFixes = ({
   )
 
   return fixes
+}
+
+let getCommentAfter = (
+  node: TSESTree.Token | TSESTree.Node,
+  source: TSESLint.SourceCode,
+): TSESTree.Comment | null => {
+  let token = source.getTokenAfter(node, {
+    filter: ({ value, type }) =>
+      type !== 'Punctuator' || ![',', ';', ':'].includes(value),
+    includeComments: true,
+  })
+
+  if (
+    (token?.type === 'Block' || token?.type === 'Line') &&
+    node.loc.end.line === token.loc.end.line
+  ) {
+    return token
+  }
+
+  return null
 }
